requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.32
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.5
python-dotenv==1.0.1
